Title: 최솟값 3개, Time: 178ms, Memory: 1MB, Status: Passed - CodeTree

+++ 240317/최솟값 3개/top-3-smallest-number.cpp
@@ -16,7 +16,7 @@ int main() {
         if(i < 3) cout << -1 << endl;
         else
         {
-            int f, s, t;
+            long long f, s, t;
             f = pq.top();
             pq.pop();
             s = pq.top();
